Move namespaced features tests to a separate file.
